WebController, testsand etc

+++ practice_tasks/kiwi_wallet_applications/open-api-kivi-wallet/src/main/kotlin/ru/oralcedbd/openapikiviwallet/api/v1/models/AccountsResponseDto.kt
@@ -3,5 +3,5 @@ package ru.oralcedbd.openapikiviwallet.api.v1.models
 data class AccountsResponseDto(
     val currencyId: Int,
     val currencyName: String,
-    val balance: Float
+    val balance: Double
 )
