+++ practice_tasks/kiwi_wallet_applications/open-api-kivi-wallet/src/main/kotlin/ru/oralcedbd/openapikiviwallet/api/v1/models/PaymentRequestDto.kt
@@ -4,6 +4,6 @@ data class PaymentRequestDto(
     val fromClientId: Long,
     val toClientId: Long,
     val currencyId: Int,
-    val summa: Float,
+    val summa: Double,
     val paymentDetail: PaymentDetailRequestDto
 )
