+++ practice_tasks/kiwi_wallet_applications/open-api-kivi-wallet/src/main/kotlin/ru/oralcedbd/openapikiviwallet/dao/WalletDao.kt
@@ -77,7 +77,7 @@ class WalletDaoImpl(
                 rs.getLong("wallet_id"),
                 rs.getLong("account_id"),
                 currencyEnumIdValueMap.toValue(rs.getInt("currency_id")),
-                rs.getFloat("balance")
+                rs.getDouble("balance")
             )
         }
     }
