+++ practice_tasks/kiwi_wallet_applications/open-api-kivi-wallet/src/main/kotlin/ru/oralcedbd/openapikiviwallet/model/Account.kt
@@ -5,5 +5,5 @@ data class Account(
     val walletId: Long,
     val accountId: Long,
     val currency: Currency,
-    val balance: Float
+    val balance: Double
 )
